pm2debug_backtrace() returns an int not a size_t
git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@17121 ee107e0b-c2fa-0310-842a-ec2723fb34f4

+++ tbx/include/tbx_macros.h
@@ -180,11 +180,11 @@
 #  define __TBX_RECORD_SOME_TRACE(array, size) pm2debug_backtrace(array, size);
 #  define __TBX_PRINT_SOME_TRACE(array, size) \
 do { \
-       size_t   i;\
+       int      i;\
        char   **strings;\
        strings = backtrace_symbols (array, size);\
 \
-       pm2debug("Obtained %zu stack frames.\n", size);\
+       pm2debug("Obtained %d stack frames.\n", size);\
 \
        for (i = 0; i < size; i++)\
           pm2debug ("%s\n", strings[i]);\
